integrated budgets into transactions

+++ src/main/java/com/cashy/cashy/budget/dto/BudgetResponseDTO.java
@@ -4,11 +4,13 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Builder
 @Getter
 @Setter
 public class BudgetResponseDTO {
     private String budgetTitle;
-    private double amountAllocated;
+    private BigDecimal amountAllocated;
     private String description;
 }
